% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaIO.R
\name{CredInt}
\alias{CredInt}
\title{CredInt}
\usage{
CredInt(x, level = 0.8)
}
\arguments{
\item{x}{Meta-analytic data}

\item{level}{alpha of credibility interval}
}
\value{
Upper and lower boundaries of the credibility interval
}
\description{
Estimates credibility interval
}

