% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaIO.R
\name{cRR}
\alias{cRR}
\title{cRR}
\usage{
cRR(x, correct_Rxx = TRUE, direct = TRUE)
}
\arguments{
\item{x}{Meta-analytic data}

\item{correct_Rxx}{If TRUE, reliability of predictor (indepedent variable) will be corrected}

\item{direct}{If TRUE, direct range restriction will be corrected. If FALSE, indirect range restriction will be corrected.}
}
\value{
Meta-analytic data corrected for range restriction (and dependent variable)
}
\description{
Conduct correction for range restriction for each effect size (i.e., each row)
based on: Hunter, Schmidt & Le (2000)
Options
1) "direct = TRUE": direct/indirect range restriction
2) "correct_Rxx = TRUE": correct/not correct for independent variable reliability
Note that Ryy: dependent variable reliability is corrected
}

