% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaIO.R
\name{cRxx}
\alias{cRxx}
\title{cRxx}
\usage{
cRxx(x)
}
\arguments{
\item{x}{Meta-analytic data}
}
\value{
Meta-analytic data corrected for independent variable reliability
}
\description{
Description: Conduct reliability correction for predictor (independent variable) for each effect size (i.e., each row)
}

