% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaIO.R
\name{cRyy}
\alias{cRyy}
\title{cRyy}
\usage{
cRyy(x)
}
\arguments{
\item{x}{Meta-analytic data}
}
\value{
Meta-analytic data corrected for dependent variable reliability
}
\description{
Description: Conduct reliability correction for criterion (dependent variable) for each effect size (i.e., each row)
}

