% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaIO.R
\name{MetaSummary}
\alias{MetaSummary}
\title{MetaSummary}
\usage{
MetaSummary(x, correct_Rxx = TRUE, correct_Ryy = TRUE, correct_RR = TRUE,
  direct = TRUE)
}
\arguments{
\item{x}{Meta-analytic data}

\item{correct_Rxx}{If TRUE, reliability of predictor (indepedent variable) will be corrected}

\item{correct_Ryy}{If TRUE, reliability of criterion (dependent variable) will be corrected}

\item{correct_RR}{If TRUE, range restriction will be corrected (see Schmidt, Hunter, Le, 2000). Note that reliability of dependent variable will be corrected during the process.}

\item{direct}{If TRUE, direct range restriction will be corrected. If FALSE, indirect range restriction will be corrected.}
}
\value{
Summary table for meta-analysis results (artifacts corrected indiviudally)
}
\description{
Summary table for meta-analysis results (artifacts corrected individually)
}
