% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaIO.R
\name{pvse2}
\alias{pvse2}
\title{pvse2}
\usage{
pvse2(x)
}
\arguments{
\item{x}{Meta-analytic data}
}
\value{
Percent variance explained by sampling error
}
\description{
Percent variance explained by sampling error
}
